requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
fastapi==0.115.10
greenlet==3.1.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
